src/agent/__init__.py
import os
from dotenv import load_dotenv
load_dotenv()

# Expose key classes for easier imports
from .models import TracerConfig, TraceResult, Entity
from .mcp_tracer import MCPTracer
from .mcp_client import MCPClient
from .mcp_http_client import MCPHTTPClient, VisualizationAPIClient
from .http_tracer import HTTPTracer
from .visualization import generate_visualization_payload
from .base_tracer import BaseTracer

