pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "amlbot_agent"
version = "0.1.1"
description = "OpenAI agent for tracing simple cases"
authors = [
    { name = "Oleg Pun", email = "oleg.pun@vareger.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic==2.12.5",
    "python-dotenv==1.1.1",
    "openai>=2.9.0",
    "click==8.2.1",
    "openai-agents>=0.6.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
amlbot-api = "agent.api:main"
amlbot-cli = "agent.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0"
]
